+++ modules/logging/core/src/sap/commerce/toolset/logging/CxLogService.kt
@@ -137,6 +137,10 @@ class CxLogService(private val project: Project, private val coroutineScope: Cor
         }
     }
 
+    fun findTemplate(templateUUID: String) = CxCustomLogTemplatesSettings.getInstance(project)
+        .templates
+        .find { it.uuid == templateUUID }
+
     private fun updateCustomLoggerTemplateInternal(template: CxCustomLogTemplateState) {
         with(CxCustomLogTemplatesSettings.getInstance(project)) {
             templates = templates.toMutableList().apply {